LayoutTests/js/dom/exceptions-thrown-in-callbacks-expected.txt

CONSOLE MESSAGE: Web SQL is deprecated. Please use IndexedDB instead.
CONSOLE MESSAGE: PASS: toString called on exception value thrown from event handler
CONSOLE MESSAGE: PASS: toString called on exception value thrown from sql transaction callback
CONSOLE MESSAGE: PASS: toString called on exception value thrown from sql error callback
CONSOLE MESSAGE: PASS: toString called on exception value thrown from timer
This test ensures that exceptions are handled correctly by the various callback mechanisms present in WebCore.
